+++ src/salespyforce/utils/tests/test_sobjects.py
@@ -7,6 +7,8 @@
 :Modified Date:  29 May 2023
 """
 
+import requests
+
 from . import resources
 
 
@@ -40,3 +42,17 @@ def test_get_and_describe_sobject():
     account_describe = sfdc_object.describe_object('Account')
     assert 'activateable' in account_describe
 
+
+def test_create_record(monkeypatch):
+    # Instantiate the core object
+    sfdc_object = resources.get_core_object()
+
+    # Overwrite the requests.post functionality with the mock_success_post() function
+    monkeypatch.setattr(requests, 'post', resources.mock_success_post)
+
+    # Perform the mock API call
+    payload = {
+        "Name": "Express Logistics and Transport"
+    }
+    response = sfdc_object.create_sobject_record('Account', payload)
+    assert 'success' in response and response.get('success') is True and 'id' in response